requirements.txt
discord.py==2.5.2
python-dotenv==1.0.0
aiohttp==3.9.1
PyYAML==6.0.1
